Area Calculation.cpp

+++ Area Calculation.cpp
@@ -1,13 +1,14 @@
 #include <iostream>
 using namespace std;
 main(){
-int A;
-int l;
-int w;
+float A;
+float l;
+float w;
 cout<<"Enter the length of the rectangle: ";
 cin>>l;
 cout<<"Enter the width of the rectangle: ";
 cin>>w;
 A = l*w;
 cout<<"The area of the rectangle is: "<<A;
+
 }
